changed copy to clipboard postition
