+++ src/SingleColor.js
@@ -1,30 +1,17 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 
-const SingleColor = ({ rgb, weight, index, hexColor }) => {
-  const [alert, setAlert] = useState(false);
+const SingleColor = ({ rgb, weight, index, hexColor, onCopy }) => {
   const bcg = rgb.join(",");
   const hexValue = `#${hexColor}`;
-  useEffect(() => {
-    const timeout = setTimeout(() => {
-      setAlert(false);
-    }, 3000);
-    return () => clearTimeout(timeout);
-  }, [alert]);
   return (
-    <>
-      <article
-        className={`color ${index > 10 && "color-light"}`}
-        style={{ backgroundColor: `rgb(${bcg})` }}
-        onClick={() => {
-          setAlert(true);
-          navigator.clipboard.writeText(hexValue);
-        }}
-      >
-        <p className="percent-value">{weight}%</p>
-        <p className="color-value">{hexValue}</p>
-        {alert && <p className="alaert">Copied to clipboard</p>}
-      </article>
-    </>
+    <article
+      className={`color ${index > 10 && "color-light"}`}
+      style={{ backgroundColor: `rgb(${bcg})` }}
+      onClick={onCopy}
+    >
+      <p className="percent-value">{weight}%</p>
+      <p className="color-value">{hexValue}</p>
+    </article>
   );
 };
